requirements.txt
annotated-types==0.7.0
bcrypt==4.2.1
blinker==1.9.0
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
Flask==3.1.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
pycparser==2.22
pydantic==2.10.2
pydantic-settings==2.6.1
pydantic_core==2.27.1
PyJWT==2.10.1
pytest==8.3.4
python-dotenv==1.0.1
six==1.16.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
webcolors==24.11.1
Werkzeug==3.1.3
